+++ EventHighway.Core/Models/Services/Orchestrations/EventArchives/V2/Exceptions/FailedEventArchiveV2OrchestrationServiceException.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections;
 using Xeptions;
 
-namespace EventHighway.Core.Models.Services.Orchestrations.EventArchives.V2
+namespace EventHighway.Core.Models.Services.Orchestrations.EventArchives.V2.Exceptions
 {
     public class FailedEventArchiveV2OrchestrationServiceException : Xeption
     {
